refactor(server): routes/ 四文件切出，index.ts 收口装配层（S2.2）
- routes/http.ts：json/readJsonBody 共享助手（避免子路由反向 import api.ts 成环）
- routes/pushRoutes.ts：/api/push/* + approval-action + approval-page
- routes/sessions.ts：/api/sessions 族 + branchOfCached（30s TTL 缓存随行）
- routes/misc.ts：fs/list、handoff、lineage、uploads、history、codex/*、config、model-names
- routes/api.ts：handleApi 聚合（原 if 链顺序保留，?? 短路）
- index.ts 收口 ~290 行：装配 init（initBackendPorts + initInbox）+ 启动守卫
  + createServer（fetch 鉴权/跨源/upgrade/静态）+ bindServer + 启动日志 + shutdown
- 路由函数体逐字搬迁；typecheck + 384 tests + build 全绿
